Add option to mirror the steering wheel
The purpose of this is basically a cheap way of getting right hand drive
cars, since many cars are left hand drive. The steering wheel location
is by far the most important thing for this: you normally can't see the
pedals, and the gauges are not operational in RBR so it doesn't really
matter if you can't see them.

This option mirrors the X coordinate of the steering wheel position
vector in the SGC file directly. It makes a copy of the SGC file before
modifying it, and alters the .ini file to point at the mirrored SGC
file.